Updates AJAX request messages on success and failure.

+++ pages/deleteUser.php
@@ -7,15 +7,18 @@
     if (isset($_SESSION["user"])) {
         $login = $_SESSION["user"];
         $account = Account::getAccountByLogin($login);
-        if ($account->isAdminAccount()){
+        if ($account->isAdminAccount()) {
+            $adminUser = User::getUserByLogin($login);
             /*TODO: check if admin wants to delete his/her own account/user -> not allowed!*/
-            if (isset($_POST["deleteUserID"])){
+            if (isset($_POST["deleteUserID"])) {
                 $deleteUserID = $_POST["deleteUserID"];
                 $userToDelete = User::getUserByID($deleteUserID);
-                Account::deleteAccountByLogin($userToDelete->login);
-                User::deleteUserByID($deleteUserID);
+                $userName = $userToDelete->getUserFullName();
+                if ($adminUser->id !== $userToDelete->id) {
+                    Account::deleteAccountByLogin($userToDelete->login);
+                    User::deleteUserByID($deleteUserID);
+                    echo "Successfully deleted user '$userName'";
+                } else echo "ERROR! You cannot delete yourself!";
             }
-        }
-        else echo "You are not an admin, sorry!";
-    }
-    else echo "session cookie 'user' not set!";
+        } else echo "<h4><error>You are not an admin, sorry!</error></h4>";
+    } else echo "<h4><error>session cookie 'user' not set!</error></h4>";
